+++ src/Settings/ModuleSettingsInterface.php
@@ -18,4 +18,8 @@ interface ModuleSettingsInterface
     public function getInvoiceNumberFormat(): string;
 
     public function getInvoiceDateFormat(): string;
+
+    public function isSendInvoiceOnUserOrderEmailActive(): bool;
+
+    public function getInvoiceInOrderEmailFilename(): string;
 }