+++ tests/Unit/Settings/ModuleSettingsTest.php
@@ -61,6 +61,19 @@ final class ModuleSettingsTest extends TestCase
         $this->assertSame($value, $sut->getInvoiceNumberFormat());
     }
 
+    public function testGetInvoiceInOrderEmailFilename(): void
+    {
+        $value = uniqid();
+
+        $mssMock = $this->createMock(ModuleSettingServiceInterface::class);
+        $mssMock->method('getString')->willReturnMap([
+            [ModuleSettings::SETTING_INVOICE_ON_ORDER_EMAIL_FILENAME, Module::MODULE_ID, new UnicodeString($value)]
+        ]);
+
+        $sut = new ModuleSettings($mssMock);
+        $this->assertSame($value, $sut->getInvoiceInOrderEmailFilename());
+    }
+
     public function testGetInvoiceDateFormat(): void
     {
         $value = uniqid();
@@ -72,6 +85,18 @@ final class ModuleSettingsTest extends TestCase
 
         $sut = new ModuleSettings($mssMock);
         $this->assertSame($value, $sut->getInvoiceDateFormat());
+    }
+
+    /** @dataProvider booleanDataProvider */
+    public function testIsSendInvoiceOnUserOrderEmailActive(bool $value): void
+    {
+        $mssMock = $this->createMock(ModuleSettingServiceInterface::class);
+        $mssMock->method('getBoolean')->willReturnMap([
+            [ModuleSettings::SETTING_SEND_INVOICE_ON_USER_ORDER_EMAIL, Module::MODULE_ID, $value]
+        ]);
+
+        $sut = new ModuleSettings($mssMock);
+        $this->assertSame($value, $sut->isSendInvoiceOnUserOrderEmailActive());
     }
 
     public function booleanDataProvider(): array