+++ views/admin_smarty/de/fainvoice_lang.php
@@ -1,5 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-require __DIR__ . "/../../admin_twig/de/fainvoice_lang.php";